cmd/tlsconfig: build tags to deprecate tls.VersionSSL30 from go1.14
